finally solved the int to char problem

+++ RasterManager/rastermanager_interface.cpp
@@ -14,7 +14,6 @@
 #include <math.h>
 #include <string>
 
-
 #include <algorithm>
 #include <iostream>
 
@@ -1019,7 +1018,7 @@ extern "C" DLL_API const char * GetReturnCodeAsString(int eErrorCode)
         return "Unspecified error.";
 
     default:
-        std::string errMsg = "Unhandled Raster Manager return code: " + std::to_string((int)eErrorCode);
+        std::string errMsg = "Unhandled Raster Manager return code: " + std::to_string((long long)eErrorCode);
         return errMsg.c_str();
     }
